edge-controller/requirements.txt
# Edge Controller Dependencies
# FastAPI Web框架
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
python-multipart>=0.0.6

# AI检测核心依赖
opencv-python>=4.8.0
numpy>=1.24.0

# HTTP客户端和异步支持
aiohttp>=3.8.0
httpx>=0.25.0
asyncio-mqtt>=0.13.0
aiofiles>=23.2.0
aiosqlite>=0.19.0

# 视频流处理
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# 配置管理
pyyaml>=6.0
python-dotenv>=1.0.0

# 图像处理
Pillow>=10.0.0

# 高级AI模型 (可选)
# mediapipe>=0.10.5
# ultralytics>=8.0.0
# torch>=2.0.0
# torchvision>=0.15.0

# 性能优化
orjson>=3.9.0

# 监控和日志
structlog>=23.1.0
prometheus-client>=0.17.0

# 开发和测试依赖
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-html>=3.1.0